fix(react-email): Watcher for subfolders (#386)

+++ packages/react-email/source/utils/watcher.ts
@@ -3,7 +3,6 @@ import {
   CLIENT_EMAILS_PATH,
   CURRENT_PATH,
   EVENT_FILE_DELETED,
-  PACKAGE_EMAILS_PATH,
   REACT_EMAIL_ROOT,
 } from './constants';
 import fs from 'fs';
@@ -21,20 +20,18 @@ export const watcher = () =>
     if (event === EVENT_FILE_DELETED) {
       const file = filename.split(path.sep);
 
-      if (file[1] === 'static') {
-        if (file[2]) {
-          await fs.promises.rm(
-            path.join(REACT_EMAIL_ROOT, 'public', 'static', file[2]),
-          );
-          return;
-        }
+      if (file[1] === 'static' && file[2]) {
+        await fs.promises.rm(
+          path.join(REACT_EMAIL_ROOT, 'public', 'static', file[2]),
+        );
+        return;
       }
 
       await fs.promises.rm(path.join(REACT_EMAIL_ROOT, filename));
     } else {
       const file = filename.split(path.sep);
 
-      if (file[1] === 'static') {
+      if (file[1] === 'static' && file[2]) {
         await copy(
           `${CLIENT_EMAILS_PATH}/static/${file[2]}`,
           `${REACT_EMAIL_ROOT}/public/static`,
@@ -42,6 +39,9 @@ export const watcher = () =>
         return;
       }
 
-      await copy(`${CLIENT_EMAILS_PATH}/${file[1]}`, PACKAGE_EMAILS_PATH);
+      await copy(
+        path.join(CURRENT_PATH, filename),
+        path.join(REACT_EMAIL_ROOT, file.slice(0, -1).join(path.sep)),
+      );
     }
   });
